addColumn, dropColumn전에 컬럼 존재 여부 판단 로직 추가

+++ src/main/java/dev/hyunlab/gravity/cmmn/service/impl/GcDbServiceImpl.java
@@ -109,33 +109,20 @@ public class GcDbServiceImpl implements GcDbService {
   @Override
   public boolean addColumn(Statement stmt, String tableName, String columnName, String dataType, String comment)
       throws SQLException {
+    if (existsColumn(stmt, tableName, columnName)) {
+      return false;
+    }
+
     GcDatabaseProductNameEnum dbProductName = GcDatabaseProductNameEnum.fromConnection(stmt.getConnection());
     switch (dbProductName) {
       case MYSQL:
       case MARIADB:
-        try (ResultSet rs = stmt.executeQuery(
-            "SELECT COUNT(*) FROM information_schema.columns WHERE table_name = '%s' AND column_name = '%s'"
-                .formatted(tableName, columnName))) {
-          rs.next();
-          if (rs.getInt(1) > 0) {
-            return false;
-          }
-        }
-
+      case POSTGRESQL:
         stmt.executeUpdate(
             "ALTER TABLE %s ADD COLUMN %s %s NULL COMMENT '%s'".formatted(tableName, columnName, dataType, comment));
-        return true;
+        break;
 
       case ORACLE:
-        try (ResultSet rs = stmt
-            .executeQuery("SELECT COUNT(*) FROM all_tab_columns WHERE table_name = '%s' AND column_name = '%s'"
-                .formatted(tableName, columnName))) {
-          rs.next();
-          if (rs.getInt(1) > 0) {
-            return false;
-          }
-        }
-
         stmt.executeUpdate("ALTER TABLE %s ADD %s %s".formatted(tableName, columnName, dataType));
         stmt.executeUpdate("COMMENT ON COLUMN %s.%s IS '%s'".formatted(tableName, columnName, comment));
         break;
@@ -149,30 +136,26 @@ public class GcDbServiceImpl implements GcDbService {
 
   @Override
   public boolean dropColumn(Statement stmt, String tableName, String columnName) throws SQLException {
+    if (!existsColumn(stmt, tableName, columnName)) {
+      return false;
+    }
+
     GcDatabaseProductNameEnum dbProductName = GcDatabaseProductNameEnum.fromConnection(stmt.getConnection());
     switch (dbProductName) {
       case MYSQL:
       case MARIADB:
         stmt.executeUpdate("ALTER TABLE %s DROP COLUMN IF EXISTS %s".formatted(tableName, columnName));
-        return true;
+        break;
 
       case ORACLE:
-        try (ResultSet rs = stmt
-            .executeQuery("SELECT COUNT(*) FROM all_tab_columns WHERE table_name = '%s' AND column_name = '%s'"
-                .formatted(tableName, columnName))) {
-          rs.next();
-          if (rs.getInt(1) == 0) {
-            return false;
-          }
-
-          stmt.executeUpdate("ALTER TABLE %s DROP COLUMN %s".formatted(tableName, columnName));
-        }
-        return true;
+        stmt.executeUpdate("ALTER TABLE %s DROP COLUMN %s".formatted(tableName, columnName));
+        break;
 
       default:
         throw new RuntimeException("Not supported db product name " + dbProductName);
     }
 
+    return true;
   }
 
   @Override
